🏗️: restructure and modularize ReplTalk
